FIX: leaked immutable stateList object when using questionData property
FIX: upload profile picture in setting page

ADD: path_id in saveDirectory state

CHANGE: variable name of image_link and image_file in questionData property

+++ app/Http/Controllers/ProfileController.php
@@ -51,7 +51,7 @@ class ProfileController extends Controller
         } else {
             $profile_pic_db_name = '';
             $query = DB::select("SELECT profile_pic from users WHERE email = :userEmail", ['userEmail' => $request->user()->email]);
-            if (file_exists(public_path() . $query[0]->profile_pic)) {
+            if (file_exists(public_path() . $query[0]->profile_pic) && $query[0]->profile_pic) {
                 unlink(public_path() . $query[0]->profile_pic);
             }
         }
